feat(web): storybook toolbar switcher for appearance theme presets
The Midnight cascade regression was only caught because one story
happened to exercise a custom theme. Promote the Theme playground's
presets to a shared module and add a Storybook toolbar (paintbrush)
that applies any preset to any story through the two real theming
paths: the clientConfig atom for antd design tokens and the Theme
component for the --theme-* CSS variables. The jotai Provider is keyed
by the selected theme so switching starts from a fresh store.
